Permissions for different sections all set up, admin customer account management created, design changes, bug fixes

+++ src/main/java/com/c3/healthapp/controllers/GripStrengthEntryController.java
@@ -32,7 +32,7 @@ public class GripStrengthEntryController {
     }
 
     @PostMapping("/target/save")
-    public ResponseEntity<Integer> saveGripTarget(@RequestBody int gripStrengthTarget) {
+    public ResponseEntity<Double> saveGripTarget(@RequestBody double gripStrengthTarget) {
         String username = SecurityContextHolder.getContext().getAuthentication()
                 .getPrincipal().toString();
         Customer customer = customerService.getCustomer(username);
